+++ modules/tealium_moments_api/example/android/app/src/main/kotlin/com/tealium/flutter/momentsapi/tealium_moments_api_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.tealium.flutter.momentsapi.tealium_moments_api_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()