+++ remotecommands/tealium_braze/example/android/app/src/main/kotlin/com/tealium/flutter/remotecommands/tealium_braze_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.tealium.flutter.remotecommands.tealium_braze_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()