+++ remotecommands/tealium_firebase/example/android/app/src/main/kotlin/com/tealium/flutter/remotecommands/firebase/tealium_firebase_example/MainActivity.kt
@@ -2,5 +2,4 @@ package com.tealium.flutter.remotecommands.firebase.tealium_firebase_example
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()