+++ tealium/example/lib/main.dart
@@ -10,8 +10,10 @@ void main() {
 }
 
 class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
@@ -23,43 +25,49 @@ class _MyAppState extends State<MyApp> {
 
   // MARK: Tealium Configuration
 
-  var config = TealiumConfig(
-      'tealiummobile',
-      'demo',
-      TealiumEnvironment.dev,
-      [Collectors.AppData, Collectors.Lifecycle],
-      [Dispatchers.RemoteCommands, Dispatchers.Collect],
-      consentPolicy: ConsentPolicy.GDPR,
-      useRemoteLibrarySettings: true,
-      batchingEnabled: false,
-      visitorServiceEnabled: true,
-      consentExpiry: ConsentExpiry(5, TimeUnit.MINUTES),
-      visitorIdentityKey: visitorIdentityKey);
+  final config = TealiumConfig(
+    'tealiummobile',
+    'demo',
+    TealiumEnvironment.dev,
+    [Collectors.AppData, Collectors.Lifecycle],
+    [Dispatchers.RemoteCommands, Dispatchers.Collect],
+    loglevel: LogLevel.DEV,
+    consentPolicy: ConsentPolicy.GDPR,
+    useRemoteLibrarySettings: true,
+    batchingEnabled: false,
+    visitorServiceEnabled: true,
+    consentExpiry: ConsentExpiry(5, TimeUnit.MINUTES),
+    visitorIdentityKey: visitorIdentityKey,
+  );
 
   @override
   void initState() {
-    // MARK: Initialize Tealium
-    Tealium.initialize(config).then((value) => {
-          developer.log('Tealium Initialized'),
-          Tealium.setConsentStatus(ConsentStatus.consented),
-          Tealium.setConsentExpiryListener(
-              () => developer.log('Consent Expired')),
-          Tealium.setVisitorServiceListener(
-              (profile) => _logVisitorProfile(profile)),
-          Tealium.getVisitorId().then((value) => _logVisitorId(value)),
-          Tealium.setVisitorIdListener((visitorId) => _logVisitorId(visitorId)),
-          Tealium.addCustomRemoteCommand('json-test',
-              (payload) => {_logRemoteCommand('JSON Test', payload)}),
-          Tealium.getFromDataLayer(visitorIdentityKey)
-              .then((value) => setState(() {
-                    userIdValue.text = value;
-                  })),
-        });
     super.initState();
+
+    // MARK: Initialize Tealium
+    Tealium.initialize(config).then((_) async {
+      developer.log('Tealium Initialized');
+      await Tealium.setConsentStatus(ConsentStatus.consented);
+      await Tealium.setConsentExpiryListener(
+        () => developer.log('Consent Expired'),
+      );
+      Tealium.setVisitorServiceListener(
+        (profile) => _logVisitorProfile(profile),
+      );
+      final visitorId = await Tealium.getVisitorId();
+      _logVisitorId(visitorId);
+      Tealium.setVisitorIdListener((visitorId) => _logVisitorId(visitorId));
+      await Tealium.addCustomRemoteCommand(
+        'json-test',
+        (payload) => _logRemoteCommand('JSON Test', payload),
+      );
+      final value = await Tealium.getFromDataLayer(visitorIdentityKey);
+      if (mounted) setState(() => userIdValue.text = value ?? '');
+    });
   }
 
   ListView _listView() {
-    return new ListView(
+    return ListView(
       scrollDirection: Axis.vertical,
       padding: EdgeInsets.all(5),
       children: <Widget>[
@@ -77,56 +85,80 @@ class _MyAppState extends State<MyApp> {
           onPressed: () => Tealium.leaveTrace(),
         ),
         TealiumButton(
-            title: 'Track Event',
-            onPressed: () => Tealium.track(
-                TealiumEvent('Some Event', {'button_click': 'test'}))),
+          title: 'Track Event',
+          onPressed: () => Tealium.track(
+            TealiumEvent('Some Event', {'button_click': 'test'}),
+          ),
+        ),
         TealiumButton(
-            title: 'Track View',
-            onPressed: () => Tealium.track(
-                TealiumView('Some View', {'screen_view': 'tester'}))),
+          title: 'Track View',
+          onPressed: () => Tealium.track(
+            TealiumView('Some View', {'screen_view': 'tester'}),
+          ),
+        ),
         TealiumButton(
-            title: 'Add Data',
-            onPressed: () =>
-                Tealium.addToDataLayer({'hello': 'world'}, Expiry.session)),
+          title: 'Add Data',
+          onPressed: () =>
+              Tealium.addToDataLayer({'hello': 'world'}, Expiry.session),
+        ),
         TealiumButton(
-            title: 'Get Data',
-            onPressed: () => Tealium.getFromDataLayer('hello').then(
-                (value) => developer.log('Value From Data Layer: $value'))),
+          title: 'Get Data',
+          onPressed: () => Tealium.getFromDataLayer('hello').then(
+            (value) => developer.log('Value From Data Layer: $value'),
+          ),
+        ),
         TealiumButton(
-            title: 'Remove Data',
-            onPressed: () => Tealium.removeFromDataLayer(['hello'])),
+          title: 'Remove Data',
+          onPressed: () => Tealium.removeFromDataLayer(['hello']),
+        ),
         TealiumButton(
-            title: 'Set Consent',
-            onPressed: () => Tealium.setConsentStatus(ConsentStatus.consented)),
+          title: 'Set Consent',
+          onPressed: () => Tealium.setConsentStatus(ConsentStatus.consented),
+        ),
         TealiumButton(
-            title: 'Get Consent',
-            onPressed: () => Tealium.getConsentStatus()
-                .then((status) => developer.log('Consent Status: $status'))),
+          title: 'Get Consent',
+          onPressed: () => Tealium.getConsentStatus().then(
+            (status) => developer.log('Consent Status: $status'),
+          ),
+        ),
         TealiumButton(
-            title: 'Set Consent Categories',
-            onPressed: () => _setRandomConsentCategories()),
+          title: 'Set Consent Categories',
+          onPressed: () => _setRandomConsentCategories(),
+        ),
         TealiumButton(
-            title: 'Get Consent Categories',
-            onPressed: () => Tealium.getConsentCategories().then((categories) =>
-                developer.log('Consent Categories: ' + categories.join(",")))),
+          title: 'Get Consent Categories',
+          onPressed: () => Tealium.getConsentCategories().then(
+            (categories) =>
+                developer.log('Consent Categories: ${categories.join(",")}'),
+          ),
+        ),
         TealiumButton(
-            title: 'Add Remote Command',
-            onPressed: () => Tealium.addCustomRemoteCommand(
-                'example', (payload) => _logRemoteCommand('Example', payload))),
+          title: 'Add Remote Command',
+          onPressed: () => Tealium.addCustomRemoteCommand(
+            'example',
+            (payload) => _logRemoteCommand('Example', payload),
+          ),
+        ),
         TealiumButton(
-            title: 'Remove Remote Command',
-            onPressed: () => Tealium.removeRemoteCommand('json-test')),
+          title: 'Remove Remote Command',
+          onPressed: () => Tealium.removeRemoteCommand('json-test'),
+        ),
         TealiumButton(
-            title: 'Get Visitor Id',
-            onPressed: () => Tealium.getVisitorId()
-                .then((visitorId) => developer.log('Visitor Id: $visitorId'))),
+          title: 'Get Visitor Id',
+          onPressed: () => Tealium.getVisitorId().then(
+            (visitorId) => developer.log('Visitor Id: $visitorId'),
+          ),
+        ),
         TealiumButton(
-            title: 'Gather Track Data',
-            onPressed: () => Tealium.gatherTrackData()
-                .then((data) => developer.log('Gather track Data: $data'))),
+          title: 'Gather Track Data',
+          onPressed: () => Tealium.gatherTrackData().then(
+            (data) => developer.log('Gather track Data: $data'),
+          ),
+        ),
         TealiumButton(
-            title: 'Terminate Tealium',
-            onPressed: () => Tealium.terminateInstance()),
+          title: 'Terminate Tealium',
+          onPressed: () => Tealium.terminateInstance(),
+        ),
         Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           mainAxisSize: MainAxisSize.min,
@@ -134,17 +166,20 @@ class _MyAppState extends State<MyApp> {
             Text('Identity:'),
             Padding(padding: EdgeInsets.all(3)),
             Expanded(
-                child: TextField(
-              controller: userIdValue,
-              autocorrect: true,
-              decoration: InputDecoration(hintText: 'Enter an Identity'),
-            ))
+              child: TextField(
+                controller: userIdValue,
+                autocorrect: true,
+                decoration: InputDecoration(hintText: 'Enter an Identity'),
+              ),
+            ),
           ],
         ),
         Padding(
           padding: EdgeInsets.symmetric(vertical: 10),
-          child: Text('Current Visitor ID: $currentVisitorId',
-              textAlign: TextAlign.center),
+          child: Text(
+            'Current Visitor ID: $currentVisitorId',
+            textAlign: TextAlign.center,
+          ),
         ),
         TealiumButton(
           title: 'Set Identity',
@@ -167,15 +202,15 @@ class _MyAppState extends State<MyApp> {
     var converted = json.decode(encodedData);
     developer.log('=========Visitor Service Response=========');
     developer
-        .log('Audiences: ' + JsonEncoder().convert(converted['audiences']));
-    developer.log('Visit Tallies: ' +
-        JsonEncoder().convert(converted['currentVisit']['tallies']));
-    developer.log('Badges: ' + JsonEncoder().convert(converted['badges']));
+        .log('Audiences: ${JsonEncoder().convert(converted['audiences'])}');
+    developer.log(
+        'Visit Tallies: ${JsonEncoder().convert(converted['currentVisit']['tallies'])}');
+    developer.log('Badges: ${JsonEncoder().convert(converted['badges'])}');
   }
 
   void _logVisitorId(String visitorId) {
     developer.log('=========Visitor Id Changed =========');
-    developer.log('VisitorId: ' + visitorId);
+    developer.log('VisitorId: $visitorId');
     setState(() {
       currentVisitorId = visitorId;
     });
@@ -186,7 +221,7 @@ class _MyAppState extends State<MyApp> {
     developer.log(JsonEncoder().convert(payload));
   }
 
-  void _setRandomConsentCategories() {
+  Future<void> _setRandomConsentCategories() async {
     List<ConsentCategories> list = [
       ConsentCategories.affiliates,
       ConsentCategories.analytics,
@@ -204,29 +239,28 @@ class _MyAppState extends State<MyApp> {
       ConsentCategories.social
     ];
     list.shuffle();
-    Tealium.setConsentCategories(list.sublist(0, 3));
+    await Tealium.setConsentCategories(list.sublist(0, 3));
   }
 
-  _joinTrace() {
-    setState(() {
-      result = traceIdValue.text;
-      Tealium.joinTrace(result);
-    });
+  Future<void> _joinTrace() async {
+    setState(() => result = traceIdValue.text);
+    await Tealium.joinTrace(result);
   }
 
-  _setIdentity() {
+  Future<void> _setIdentity() async {
     var identity = userIdValue.text;
-    Tealium.addToDataLayer({visitorIdentityKey: identity}, Expiry.forever);
+    await Tealium.addToDataLayer({visitorIdentityKey: identity}, Expiry.forever);
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       home: Scaffold(
-          appBar: AppBar(
-            title: const Text('TealiumFlutterPluginExample'),
-          ),
-          body: _listView()),
+        appBar: AppBar(
+          title: const Text('TealiumFlutterPluginExample'),
+        ),
+        body: _listView(),
+      ),
     );
   }
 }